requirements.txt
alembic==1.0.11
Babel==2.7.0
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.7
Werkzeug==0.15.5
WTForms==2.2.1
